requirements.txt
feedparser==6.0.10
confluent-kafka==2.3.0
python-dotenv==1.0.0
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.38.1
fastapi==0.104.1
uvicorn[standard]==0.24.0